+++ modules/identity/src/Volo.Abp.Identity.Domain/Volo/Abp/Identity/AbpDataProtectionTokenProviderOptions.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Volo.Abp.Identity.AspNetCore;
+namespace Volo.Abp.Identity;
 
 /// <summary>
 /// Stands in for ASP.NET Core's <c>DataProtectionTokenProviderOptions</c>, which is only available
