+++ modules/identity/src/Volo.Abp.Identity.Domain/Volo/Abp/Identity/AbpDefaultTokenProviderOptions.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.AspNetCore.Identity;
 
-namespace Volo.Abp.Identity.AspNetCore;
+namespace Volo.Abp.Identity;
 
 public class AbpDefaultTokenProviderOptions : AbpDataProtectionTokenProviderOptions
 {
