+++ modules/identity/src/Volo.Abp.Identity.Domain/Volo/Abp/Identity/AbpEmailTwoFactorTokenProviderOptions.cs
@@ -1,4 +1,4 @@
-namespace Volo.Abp.Identity.AspNetCore;
+namespace Volo.Abp.Identity;
 
 public class AbpEmailTwoFactorTokenProviderOptions : AbpTwoFactorTokenProviderOptions
 {
