+++ modules/identity/src/Volo.Abp.Identity.Domain/Volo/Abp/Identity/AbpTwoFactorTokenProviderOptions.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Volo.Abp.Identity.AspNetCore;
+namespace Volo.Abp.Identity;
 
 public abstract class AbpTwoFactorTokenProviderOptions
 {
