+++ modules/identity/src/Volo.Abp.Identity.Domain/Volo/Abp/Identity/IdentityUserManagerSingleActiveTokenExtensions.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 
-namespace Volo.Abp.Identity.AspNetCore;
+namespace Volo.Abp.Identity;
 
 /// <summary>
 /// Provides extension methods on <see cref="IdentityUserManager"/> for invalidating
