+++ modules/identity/test/Volo.Abp.Identity.TestBase/Volo/Abp/Identity/CrossHostTokenProvider_Tests.cs
@@ -13,7 +13,6 @@ using Shouldly;
 using Volo.Abp.Autofac;
 using Volo.Abp.Data;
 using Volo.Abp.Domain.Repositories;
-using Volo.Abp.Identity.AspNetCore;
 using Volo.Abp.Modularity;
 using Volo.Abp.MultiTenancy;
 using Volo.Abp.Uow;
